market_data_service/requirements.txt
asyncio==3.4.3
websockets==12.0
aiohttp==3.9.1
asyncpg==0.29.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
aio-pika==9.3.1
pydantic==2.5.0
pydantic-settings==2.1.0
prometheus-client==0.19.0
structlog==23.2.0
uvloop==0.19.0
python-binance==1.0.19
numpy==1.26.2
pandas==2.1.4
fastapi==0.104.1
uvicorn==0.24.0
textblob==0.17.1
nltk==3.8.1
requests==2.31.0
yfinance==0.2.28
# Technical Analysis Library
ta==0.10.2
# For Parquet export (backtesting)
pyarrow==14.0.1
fastparquet==2023.10.1
# For scheduled tasks
schedule==1.2.0
# For statistical correlation analysis
scipy==1.11.4
# For Reddit API access
praw==7.7.1
# For Telegram API access (optional)
telethon==1.35.0